% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatFromDF.R
\name{adjacencyMatFromDF}
\alias{adjacencyMatFromDF}
\title{Get an adjacency matrix from an edge data frame}
\usage{
adjacencyMatFromDF(edges, nodes = NULL)
}
\arguments{
\item{edges}{a data frame with columns a, b, and (optionally) weight}

\item{nodes}{an array of the unique node IDs used in a, b (inferred if not provided)}
}
\value{
an adjacency matrix
}
\description{
Get an adjacency matrix from an edge data frame.
If no weights are provided all weights are set to 1
}
