% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueNodesFromEdgesDF.R
\name{getUniqueNodesFromEdgesDF}
\alias{getUniqueNodesFromEdgesDF}
\title{Get unique node array from an edges dataframe}
\usage{
getUniqueNodesFromEdgesDF(edges)
}
\arguments{
\item{edges}{data frame with columns a, b}
}
\value{
a sorted array of the unique elements appearing in columns a, b
}
\description{
Get unique node array from an edges dataframe.
}
