% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCommunitiesASLP.R
\name{inferCommunitiesASLP}
\alias{inferCommunitiesASLP}
\title{Community detection, asynchronous label propagation}
\usage{
inferCommunitiesASLP(edges)
}
\arguments{
\item{unique_edges}{a data frame with columns a, b, weight representing the connections between nodes.
We assume undirected graph, and therefore b < a.}
}
\value{
a list of two vectors,
node--node id (taken from a, b of input) and
label--unique cluster/community ID.
}
\description{
An implementation of community detection by label propagation in an undirected weighted graph based on
Raghavan, Albert, Kumara. Phys Rev E 76, 036106 (2007)
}
\details{
This is fully asynchronous -- labels are updated one at a time in a random order.
}
\author{
Yuriy Sverchkov
}
