% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCommunitiesGreedily.R
\name{inferCommunitiesGreedily}
\alias{inferCommunitiesGreedily}
\title{Community detection, greedy algorithm}
\usage{
inferCommunitiesGreedily(similarities, save.file = NULL,
  full_trajectory = F, simplify = T, loop_limit = Inf)
}
\arguments{
\item{similarities}{a data frame with columns a, b, similarity representing the similarities between nodes.
We assume undirected graph and b < a.}
}
\value{
a data frame with two columns, n - node id (taken from a, b of input) and cluster - unique cluster ID.
}
\description{
A naive implementation of community detection in a fully-connected weighted graph based on
Newman ME, Girvan M. Finding and evaluating community structure in networks.
Physical review E. 2004 Feb 26;69(2):026113.
}
\author{
Yuriy Sverchkov
}
