% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCommunitiesLP.R
\name{inferCommunitiesLP}
\alias{inferCommunitiesLP}
\title{Community detection, label propagation}
\usage{
inferCommunitiesLP(unique_edges, async_prop = 0.5, check_unique = F)
}
\arguments{
\item{unique_edges}{a data frame with columns a, b, weight representing the connections between nodes.
We assume undirected graph, and therefore b < a.}

\item{async_prop}{proportion of nodes to update before synchronous update}

\item{check_unique}{whether to check edges data frame for uniqueness}
}
\value{
a data frame with two columns,
node--node id (taken from a, b of input) and
label--unique cluster/community ID.
}
\description{
An implementation of community detection by label propagation in an undirected weighted graph based on
Raghavan, Albert, Kumara. Phys Rev E 76, 036106 (2007)
}
\author{
Yuriy Sverchkov
}
