% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initCommunitySearch.R
\name{initCommunitySearch}
\alias{initCommunitySearch}
\title{Initialize community search objects}
\usage{
initCommunitySearch(similarities)
}
\arguments{
\item{similarities}{see [inferCommunitiesGreedily]}
}
\description{
Initialize the communities and membership tables
}
\author{
Yuriy Sverchkov
}
