% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCommunitiesLP.R
\name{voteForLabelPropagation}
\alias{voteForLabelPropagation}
\title{Helper, implements voting part of label propagation}
\usage{
voteForLabelPropagation(edges, nodes)
}
\arguments{
\item{edges}{edges data frame}

\item{nodes}{nodes data frame}
}
\description{
Helper implements voting part of label propagation
}
